major project refactoring
