+++ src/main/java/edu/baylor/GroupFive/database/controllers/AccountController.java
@@ -1,8 +1,8 @@
-package edu.baylor.GroupFive.controllers;
+package edu.baylor.GroupFive.database.controllers;
 
 import edu.baylor.GroupFive.models.User;
-import edu.baylor.GroupFive.services.AccountService;
-import edu.baylor.GroupFive.services.LoginService;
+import edu.baylor.GroupFive.database.services.AccountService;
+import edu.baylor.GroupFive.database.services.LoginService;
 
 // Assume a User is logged in when calling these functions
 // i.e. Don't need to read pswd
