+++ src/main/java/edu/baylor/GroupFive/database/services/LoginService.java
@@ -1,11 +1,11 @@
-package edu.baylor.GroupFive.services;
+package edu.baylor.GroupFive.database.services;
 
-import edu.baylor.GroupFive.database.userDAO.UserDatabaseConnection;
+import edu.baylor.GroupFive.database.daos.UserDAO;
 import edu.baylor.GroupFive.models.User;
 
 public class LoginService {
     public static User login(String username, String password) {
-        UserDatabaseConnection conn = new UserDatabaseConnection();
+        UserDAO conn = new UserDAO();
         return conn.getUser(username);
     }
 }
