+++ src/main/java/edu/baylor/GroupFive/ui/modifyReservation/ModifyReservationPanel.java
@@ -13,8 +13,8 @@ import javax.swing.JLabel;
 import javax.swing.JPanel;
 import javax.swing.JTextField;
 
-import edu.baylor.GroupFive.CoreUtils;
-import edu.baylor.GroupFive.controllers.ReservationController;
+import edu.baylor.GroupFive.util.CoreUtils;
+import edu.baylor.GroupFive.database.controllers.ReservationController;
 import edu.baylor.GroupFive.models.Reservation;
 import edu.baylor.GroupFive.ui.utils.DatePanel;
 import edu.baylor.GroupFive.ui.utils.Page;
