+++ src/main/java/edu/baylor/GroupFive/util/exceptions/BadConnectionException.java
@@ -1,4 +1,4 @@
-package edu.baylor.GroupFive.exceptions;
+package edu.baylor.GroupFive.util.exceptions;
 
 // Thrown when a connection cannot be made to the database
 public class BadConnectionException extends HotelProjectException {
