+++ src/main/java/edu/baylor/GroupFive/util/exceptions/InvalidCredentialsException.java
@@ -1,4 +1,4 @@
-package edu.baylor.GroupFive.exceptions;
+package edu.baylor.GroupFive.util.exceptions;
 
 // Thrown if invalid credentials are provided when verifying a user
 public class InvalidCredentialsException extends HotelProjectException {
