+++ src/main/java/edu/baylor/GroupFive/util/exceptions/SkillIssueException.java
@@ -1,4 +1,4 @@
-package edu.baylor.GroupFive.exceptions;
+package edu.baylor.GroupFive.util.exceptions;
 
 // Throw if just bad at coding
 public class SkillIssueException extends HotelProjectException {
