+++ src/main/java/edu/baylor/GroupFive/util/exceptions/UnknownException.java
@@ -1,4 +1,4 @@
-package edu.baylor.GroupFive.exceptions;
+package edu.baylor.GroupFive.util.exceptions;
 
 public class UnknownException extends HotelProjectException {
     public UnknownException() {
